scripts: sign_spl_fit: fix an encryption issue with CST 3.3.2
The uboot NXP security guide in paragraph 1.7.1 indicates to copy the Nonce/MAC from the binary CSF relating
to encryption into the binary CSF relating to signature. It is done that way because the first CSF performs a
signature on the non-ciphered text, while the second performs the useful signature, the one on the ciphertext.

The nonce offset was calculated starting from the end of the binary CSF, substracting 36 bytes, which is the
size of the Nonce/MAC. In fact it assumes that the Nonce/MAC is the last element in the binary CSF. For CST
3.3.2 this is no longer true. CST 3.3.2 writes at the end of the binary CSF the "AuthenticateCSF data", i.e.
the signature, and not the Nonce/MAC. This makes the procedures outlined in the guides unenforceable.

The fix is to scan the CSF binary looking for the Nonce/MAC HAB TAG to calculate the correct offset in each
iteration. It must be done in that way because the "AuthenticateCSF data" size is variable and we can no longer
rely on a fixed size/offset.

While on it, make use of the "conv=notrunc" flag when copying the Nonce/MAC using "dd" to avoid final file to
be truncated at the end of the Nonce/MAC.

CST 3.3.1
=========

            ------------ +-----------------------------+ <-- *start
            ^            |                             |
            |            |                             |
  Command   |            |            CSF              |
  Sequence  |            |            Data             |
  File      |            |                             |
            |            |                             |
  (CSF)     |    ------- +-----------------------------+ <-- *total size
            |         ^  |                             |      minus 36
            |    36   |  |                             |
            |   bytes |  |         Nonce/MAC           |
            v         v  |                             |
            ------------ +-----------------------------+

CST 3.3.2
=========

            ------------ +-----------------------------+ <-- *start
            ^            |                             |
            |            |                             |
  Command   |            |            CSF              |
  Sequence  |            |            Data             |
  File      |            |                             |
            |            |                             |
  (CSF)     |    ------- +-----------------------------+ <-- *unknwon
            |         ^  |                             |      offset
            |    36   |  |                             |
            |   bytes |  |         Nonce/MAC           |
            |         v  |                             |
            |     ------ +-----------------------------+
            |         ^  |                             |
            | Unknown |  |    AuthenticateCSF data     |
            |    size |  |         (Signature)         |
            v         v  |                             |
            ------------ +-----------------------------+

Signed-off-by: David Escalona <david.escalona@digi.com>