ccimx93: configure u-boot environment in emmc regardless of boot media
Current build configuration supports dual bootloader with the
environment (primary and redundant) at the end of the hardware bootX
partitions respectively.

That configuration is not compatible with having the environment on the
SD itself when booting from an SD card. You can do that, but then you
need to change the compilation configuration and build a different
u-boot.

Thus to allow the SD card booting without errors, just configure the
environment to be in the EMMC regardless of the boot media (EMMC/SD).

https://onedigi.atlassian.net/browse/DUB-1067

Signed-off-by: Javier Viguera <javier.viguera@digi.com>