scripts: sign_spl_ahab: fix parsing ATF container offset
This sign script parses the ATF container offset from the log file
mkimage generates when building the bootloader.

The ccimx93 uses a new version of mkimage, which produces a slightly
different output log. This was making the parsing to fail.

The commit makes the regular expression more robust, so it works with old
and new versions of mkimage.

https://onedigi.atlassian.net/browse/DUB-1068

Signed-off-by: Javier Viguera <javier.viguera@digi.com>