f_fastboot: update partition table when switching fastboot_dev
Our DEY UUU installation scripts switch fastboot_dev to sata and back to mmc
in order to update fastboot's info regarding the eMMC partition table. This is
useful when switching from a single-boot system to a dualboot one and
vice-versa, as the partition table gets changed during the process.

However, the FSL fastboot implementation in U-Boot v2017.03 doesn't have this
logic, so the partition info only gets loaded once when U-Boot starts, it
doesn't get updated when repartitioning the eMMC, and the script fails.

To fix this, port two additional steps from the v2020.04 implementation of FSL
fastboot:

    * Whenever a U-Boot command is run via UUU, call fastboot_setup() in case
      the command affects the fastboot environment.
    * In fastboot_setup(), only load the partition table if the fastboot dev
      info has been switched. Otherwise, it will load the parititon table with
      every command, adding unnecessary overhead.

This makes it so fastboot's partition table is up to date if the eMMC is
repartitioned.

https://onedigi.atlassian.net/browse/DEL-8858

Signed-off-by: Gabriel Valcazar <gabriel.valcazar@digi.com>