ccimx93: enable trustfence update support
The trustfence update command allows programming an encrypted bootloader.

Notice that the "Image Indexes" value in the CSF template changed to
encrypt all the images within any AHAB container. The ccimx93 needs this
because the second container only has one image (the SPL image), and
the CST fails otherwise because at least one image per container has to
be encrypted.

https://onedigi.atlassian.net/browse/DEL-8945

Signed-off-by: Javier Viguera <javier.viguera@digi.com>