dboot: reimplement function to compose the FIT boot command
The "get_cfgs_cmd_from_fit" function parses the FIT image on memory to
extract the different configurations and compose the boot command. For
an encrypted image, this is no longer possible. You would need to decrypt
the FIT image before, which is time-consuming.

Instead, reimplement the function to compose the boot command using a
compile-time (but runtime overrideable) prefix and the contents of the
"overlays" environment variable. Rename the function as well, as it's
not reading the FIT image in memory anymore.

https://onedigi.atlassian.net/browse/DEL-8945

Signed-off-by: Javier Viguera <javier.viguera@digi.com>