trustfence: support encrypted FIT boot artifacts authentication
We already supported signed FIT images. This commit implements support
to authenticate signed+encrypted FIT-based boot artifacts.

Notice that for encrypted images, unlike the signed-only ones, we need to
decrypt the FIT image to read the boot script in the source command. For
signed-only images, the payload is readable, and just adjusting the
address in memory is enough to read the boot script.

https://onedigi.atlassian.net/browse/DEL-8945

Signed-off-by: Javier Viguera <javier.viguera@digi.com>