ccimx8, ccimx9: disable CONFIG_IMX_OPTEE
NXP uses this option only on imx6/7-based devices. It is disabled for
imx8/9-based devices. We changed the behavior to amend the device tree
with optee information only when that option is enabled, but that does
not match what NXP does for similar platforms.

Revert to NXP code for easier maintenance.

Signed-off-by: Javier Viguera <javier.viguera@digi.com>