ccimx8x_sbc_pro: fix trustfence environment encryption
The ccimx8x implements CAAM-based environment encryption. For this to
work, u-boot needs to initialize the CAAM early enough to decrypt the
environment in the media and load it to memory.

The previous code tried initializing the CAAM using the old, not
DM-based code. This commit initializes the CAAM properly at a point
(board_early_init_r function) early enough for u-boot to decrypt the
environment.

https://onedigi.atlassian.net/browse/DUB-1080

Signed-off-by: Javier Viguera <javier.viguera@digi.com>