ccimx6ul: Add MMDC_MSCR check to DCD table
Some units may not boot through USB when the DCD table is sent to be
written in a row by the ROM loader. That happens because the MMDC_MSCR
register, which is used to issue commands to the external DDR, is not
ready to accept a new command when the final write operation is performed.

To avoid that, ensure MMDC_MSCR register CON_ACK flag is set to 1,
indicating that configuration of MMDC registers is permitted, before
attempting the final configuration to put the DDR into Normal operation
mode.

https://onedigi.atlassian.net/browse/DEL-8779

Signed-off-by: Gonzalo Ruiz <Gonzalo.Ruiz@digi.com>