scripts: sign_spl_fit.sh: make mkimage log parsing more robust
Our parsing logic assumes that there is only one match per pattern. This was
true until we applied the CVE-2023-39902 fixes to imx-mkimage, which make it so
there are now two matching lines for the " csf_off" pattern, resulting in a
non-numerical value which causes the script to silently fail when inserting the
SPL CSF into the imx-boot image.

Make our parsing logic more robust by adding a "line start" to each pattern.

https://onedigi.atlassian.net/browse/DUB-1081

Signed-off-by: Gabriel Valcazar <gabriel.valcazar@digi.com>