ccmp1: Adjust FASTBOOT_BUF_SIZE
Current buffer (240MB) was too big for CCMP13 standard variant (256MB)
and too small for CCMP15 standard variant (512MB).

Define a buffer size equivalent to the whole DDR. It will rely on the
user to control the size of the files being downloaded through the
fastboot/UUU tool:

FASTBOOT_BUF_SIZE
- MP15: 1GB (maximum DDR in all variants)
- MP13: 256MB

The maximum file size that can be transferred is dependent on the memory
reserved regions:

Memory reserved regions on MP15:
- 32MB for the OP-TEE
- 64MB for the GPU
- 38MB for other purposes

Memory reserved regions on MP13:
- 32MB for the OP-TEE
- 16MB for the OP-TEE Framebuffer
- 36MB for other purposes

Maximum file size:
- MP13 (256MB): 172MB
- MP15 (512MB): 378MB
- MP15 (1GB):   890MB

These numbers consider the fastboot buffer to start at address 0xC0000000,
the beginning of the DDR. If the download is performed on a buffer
starting in loadaddr (0xC2000000) instead, consider the maximum file size
has to be reduced 32MB more.

Signed-off-by: Gonzalo Ruiz <Gonzalo.Ruiz@digi.com>