ccimx8mn: set soc_type to 'imx8mnsl' for Solo Lite CPU variants
Define soc_type with a different value when the i.MX8MN CPU variant is a
Quad Core or a Solo Lite. This will help customize the software if there
is any difference between the variants.

This was done on v2018.03/maint and was lost after moving to a newer
U-Boot version.

Signed-off-by: Gonzalo Ruiz <Gonzalo.Ruiz@digi.com>