ccimx8mm: default to 4GB DDR config when HWID unprogrammed
The differences in geometry between a 4GB LPDDR4 (2 chip selects) and
the 2GB/1GB/512MB configurations (1 chip select) makes it impossible to
select a default DDR config that can boot on every variant.

Loading a 1GB configuration on a 4GB LPDDR4 fails randomly, while
loading a 4GB configuration on a 1GB/2GB LPDDR4 fails the training.

As a result, we assume by default the DDR will have 2 chip selects and
initialize the minimum size for that geometry (4GB). In case of DDR init
failure we fall back to assuming the DDR has 1 chip select and we
initialize the minimum size for that geometry (1GB).

Once initialized, in U-Boot just map 1GB in every case.

https://onedigi.atlassian.net/browse/DEL-9494

Signed-off-by: Gonzalo Ruiz <Gonzalo.Ruiz@digi.com>