trustfence: get rid of the public get_dek_blob_size function
Trustfence command was only using that function once, and that can be
easily replaced by a DEK_BLOB_SIZE macro that calculates the blob size
out of the plain DEK.

Make get_dek_blob_size static where used internally, and remove it for
the ccimx6 where it is not used anymore. Also for the ccimx8m, rewrite
it using a DEK blob header struct.

https://onedigi.atlassian.net/browse/DUB-1099

Signed-off-by: Javier Viguera <javier.viguera@digi.com>