ARM: dts: ccmp15: add missing optee and scmi nodes
For U-Boot to access secure peripherals (such as the BSEC for accessing
OTP bits), it needs to communicate with OP-TEE.
On the MP15 such nodes only existed in specific DTSI files (with -scmi on
the filename).
Including these DTSI files has side effects, so just bring the required
nodes into the ccmp15 SOM dtsi.
This unblocks the access of U-Boot to BSEC via SMC.

Signed-off-by: Hector Palacios <hector.palacios@digi.com>