crypto: fsl: RNG workaround for i.MX6UL
The i.MX6UL processor, under some conditions, could fail the validation
of the entropy generated by the TRNG. This is caused by a problem on the
RNG health test. The workaround performs the validation using a software
routine, running from the OCRAM, instead of using the HW support. The
bootloader loads the firmware to do the validation in the OCRAM memory
if the RNG workaround is enabled.

Signed-off-by: yanxing cui <yanxing.cui@nxp.com>
Signed-off-by: Pedro Perez de Heredia <pedro.perez@digi.com>
(cherry picked from commit 5d3666873fc48cb9f2104a66ad4901692fdd298c)