Revert "cmd: nand/sf: update legacy mtd code"
The change returns the index of the partition which is later
used by NAND code to set the active MTD device.
This is supposed to be just the NAND (0), not the partition,
and is constrained to be within the limit marked by
SYS_MAX_NAND_DEVICE. This resulted in error when doing nand
operations, such as 'nand erase.part <part>'.

This reverts commit d2bf319761140419fba4264fbf4e1897cc015a95.

Signed-off-by: Hector Palacios <hector.palacios@digi.com>