Revert "trustfence: env: remove redundant sec initialization"
On the ccimx6ul, there's a corner case that prevents the system from booting.
The conditions are:

    * Secure boot is enabled and the bootscript is signed
    * Environment encryption is enabled
    * Dualboot is enabled
    * The bootscript's dualboot logic triggers an "env save" to sanitize boot
      parameters

When this happens, the CAAM is used to encrypt the environment shortly after
it's used to authenticate the bootscript, which triggers a blob encapsulation
error. The CAAM clk and the sec subsystem have to be initialized between these
two operations (authentication and encryption) in order to avoid the error.

This reverts commit 833a7732fa3c8c3eef4518795445ce9e5adc0b0e.

https://onedigi.atlassian.net/browse/DEL-9591

Signed-off-by: Gabriel Valcazar <gabriel.valcazar@digi.com>