sign_spl_ahab: make image indexes mask configurable for encryption
Replace the hardcoded "Image Indexes" value in the AHAB encryption CSF
template with a configurable placeholder.

Update the "sign_spl_ahab.sh" script to use a default value of 0xFFFFFFFE
for ConnectCore 8X platforms, and 0xFFFFFFFF for ConnectCore 9 platforms.

The 0xFFFFFFFE mask is required on CC8X platforms to prevent encryption of
the SCU firmware. Encrypting the SCU image on i.MX8X B0 silicon results
in a boot failure. On the other hand, for CC9 platforms, the SPL AHAB
container has one image only (the SPL), so the mask needs to be 0xFFFFFFFF.

https://onedigi.atlassian.net/browse/DUB-1104

Signed-off-by: Javier Viguera <javier.viguera@digi.com>