autoboot: increase in 1 second the timeout to input a password
The time to input the password to recover the silent console is limited
to the bootdelay. If this time is small (1 second), it's pretty unlikely
that you can enter the password.
This commit increases the time in one second on each key press if the
remaining time is less than a second (or has slightly expired), with a
limit of 20 chars to avoid key presses from stopping auto-boot.

Signed-off-by: Hector Palacios <hector.palacios@digi.com>

https://onedigi.atlassian.net/browse/DUB-1107