autoboot: add 1 second to user input password
The time to input a password is limited and with a silent console
it's difficult to know when to start typing it. Typing too early
results in missin characters and typing too late results in timeout.
Add 1 second to the total delay (bootdelay) to give a little more
margin to the user input.

Signed-off-by: Hector Palacios <hector.palacios@digi.com>

https://onedigi.atlassian.net/browse/DUB-1107