autoboot: don't stop on any keypress if console is silent
If console is silent but no special password is set for
recovery, any key can stop the autoboot process and recover
from silent console. This happens, for instance, when
configuring a GPIO as enabler for the silent console.

Check if the console is silent before accepting a key to
stop the autoboot.

Signed-off-by: Hector Palacios <hector.palacios@digi.com>

https://onedigi.atlassian.net/browse/DUB-1108