ccmp1/ccmp2: set source_fit_script as a local variable
The loadscript was setting 'source_fit_script' as an env
variable when 'dboot_kernel_var' is set to 'fitimage', however
nobody's unsetting the variable for the regular case.
In normal conditions, the boot script boots the operating
system and the variable is lost because the environment
is not saved during the boot process, but if any of the commands
in the boot script fails and the control is returned to the
user, the environment may be saved with this variable.
In that case, returning to boot a regular system (not fitimage)
would fail cause the bootcmd would use the value in the
variable.
A local variable avoids this corner case.

Signed-off-by: Hector Palacios <hector.palacios@digi.com>