ccmp2: disable CoreSight and debug features by default
Coresight IPs and related debug features are recommended to be disabled when
the device is in secure closed state. To avoid maintaining separate Device Tree
files for open and closed configurations, this commit disables all such
features by default.

CoreSight components are integrated into the processor to enable system-level
debugging and tracing, offering real-time visibility that is valuable during
development and validation of complex embedded systems. However, they are not
meant to remain active by default on production or security-sensitive devices.

https://onedigi.atlassian.net/browse/DEL-9734

Signed-off-by: Arturo Buzarra <arturo.buzarra@digi.com>