imx95: weaken SoC env_get_location() to use board implementation
The SoC implementation sets the location on the boot device and
defaults to NOWHERE or UNKNOWN (for instance when you boot from USB).
We want the location of the environment to be always the MMC
regardless of the boot device. This helps saving the environment
when booting from USB in manufacturing mode.

Signed-off-by: Hector Palacios <hector.palacios@digi.com>