board: digi: ccimx6sbc: drop legacy Ethernet PHY initialization
The ccimx6sbc board code was manually powering up and configuring the
Ethernet PHY during board init. This is no longer required since the PHY
is fully described in the device tree and handled by the DM Ethernet and
MDIO subsystems.

Also fix wrong orientation in dtbs of reset line.

https://onedigi.atlassian.net/browse/DEL-9756

Signed-off-by: Francisco Gil <francisco.gilmartinez@digi.com>