scripts: sign_spl_fit.sh: use mac_dump tool to obtain nonce/MAC parameters
We were using an ad-hoc method with a hardcoded nonce/MAC length, but NXP
provides a tool for this since CST v3.4.1, so use it instead.

CSF size hasn't been used in nonce/MAC offset calculations for some time, so
there's no need to calculate it or use it anymore.

https://onedigi.atlassian.net/browse/DEL-9334

Signed-off-by: Gabriel Valcazar <gabriel.valcazar@digi.com>