update: strictly check 'env_get_yesno()' returns true
The function 'env_get_yesno()' may return a negative error code.
When used on an 'if' conditional, we must not use it as a boolean but
check instead the exact return value.
On a corner case where the DEY installers are run on a unit where the
'dualboot' variable does not exist, this condition was met, when it
shouldn't have. As a result, the update command tried to operate on a
partition 'linux_a' (for a dual boot system) when it should operate over
partition 'linux' (for a single boot system).

Signed-off-by: Hector Palacios <hector.palacios@digi.com>

https://onedigi.atlassian.net/browse/DEL-9883