ccimx95: limit update operations to first memory bank
On every update operation where there is enough RAM space, the written
firmware is read on a second location in order to verify its integrity.

That second location is obtained dynamically considering the whole DDR
size, but for CC95 that lied in the second bank (addresses 0x100000000
to 0x27FFFFFFF), which presented an access error when the MMC was read:

    Cannot use 64 bit addresses with SDMA

First bank has 1.75GB of free space (addresses 0x90000000 to
0xFFFFFFFF), so limit the update operations to take place just in the
first bank.

https://onedigi.atlassian.net/browse/DEL-9885

Signed-off-by: Gonzalo Ruiz <Gonzalo.Ruiz@digi.com>