ARM64: dts: ccmp25-dvk: set LDO6/7 voltages
LDO6 and LDO7 are not used inside the SOM. Their voltages
were originally constrained on OPTEE device tree to match
their usage on the DVK, but this made it more difficult
to set different voltage on custom boards.
The constraints have been removed from OPTEE device tree and
now the LDOs can operate on their full range (0.9 ~ 4.0V).

This commit specifically sets the voltages of these LDOs
for the DVK:
 - LDO6 is PCIe 1.5V
 - LDO7 is HDMI 1.8V

NOTE: these regulators are not currently used/enabled in U-Boot.

Signed-off-by: Hector Palacios <hector.palacios@digi.com>

https://onedigi.atlassian.net/browse/DEL-9909